new apps/ directory with expenses app
