+++ src/apps/README.md
@@ -1,0 +1,24 @@
+# Apps
+
+This folder contains all the apps of Open Collective.
+It's one step towards making the code base more modular to eventually enable third party developers to add more apps.
+
+- Expenses:
+File expenses, approve them, show transactions
+TODO: Move transactions to their own app
+
+- Updates
+Create and publish updates to your members
+TODO: Move into its own app
+
+- Team
+Add, remove people to your team
+TODO: Move into its own app
+
+- Events
+Create, edit events with their own ledger
+TODO: Move into its own app
+
+- Goals
+Add, edit fund raising goals
+TODO: Move into its own app
