+++ src/apps/expenses/components/ExpensesStatsWithData.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import withIntl from '../lib/withIntl';
 import { FormattedMessage } from 'react-intl';
 import { graphql } from 'react-apollo'
 import gql from 'graphql-tag'
 import { get } from 'lodash';
-import Link from './Link';
-import Currency from './Currency';
+
+import withIntl from '../../../lib/withIntl';
+import Link from '../../../components/Link';
+import Currency from '../../../components/Currency';
+
 
 class ExpensesStatsWithData extends React.Component {
