+++ src/apps/expenses/components/TransactionDetails.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { defineMessages, FormattedNumber, FormattedMessage } from 'react-intl';
-import { imagePreview, capitalize } from '../lib/utils';
-import withIntl from '../lib/withIntl';
+import { imagePreview, capitalize } from '../../../lib/utils';
+import withIntl from '../../../lib/withIntl';
+
 import RefundTransactionBtn from './RefundTransactionBtn';
+
 
 class TransactionDetails extends React.Component {
