+++ src/pages/transactions.js
@@ -1,15 +1,18 @@
 import React from 'react';
+import withData from '../lib/withData';
+import withIntl from '../lib/withIntl';
+
+import { addCollectiveCoverData, addGetLoggedInUserFunction } from '../graphql/queries';
 import Header from '../components/Header';
 import Body from '../components/Body';
 import Footer from '../components/Footer';
 import CollectiveCover from '../components/CollectiveCover';
-import { addCollectiveCoverData, addGetLoggedInUserFunction } from '../graphql/queries';
 import Loading from '../components/Loading';
 import NotFound from '../components/NotFoundPage';
 import ErrorPage from '../components/ErrorPage';
-import withData from '../lib/withData';
-import withIntl from '../lib/withIntl';
-import TransactionsWithData from '../components/TransactionsWithData';
+
+import TransactionsWithData from '../apps/expenses/components/TransactionsWithData';
+
 
 class TransactionsPage extends React.Component {
